SendBufferSize is not valid in a VirtualHost, so make Apache say so.
Note that allowing this to be set on a per socket (ie. per Listen
directive) basis could be useful for some, but isn't currently
implemented.  Also note that the code as written may fail to work
correctly on some systems that don't copy the socket settings fully
on an accept().

Reviewed by:	Randy Terbush, Dean Gaudet
Submitted by:
Obtained from:


git-svn-id: https://svn.eu.apache.org/repos/asf/httpd/httpd/trunk@78281 13f79535-47bb-0310-9956-ffa450edef68