Make our nitpicking and complaining guys happy:
Added the mod_rewrite `URL Rewriting Guide' to the online documentation
(htdocs/manual/misc/rewriteguide.html). This paper provides a large
collection of practical solutions to URL based problems a webmaster is
often confronted with.

This version of the text was translated from my WML source on my website
and my old official version is now discarded. So, as it was requested,
this can be treated as an official donation of this text to the ASF.
This way the ASF is now the official owner of this text.

                               - - -

<IRONIC>
Be happy and give Ken and Jim the credit for achieving this by being
such sensitive and friendly to other developers like me and always
reminding us that it is such contemptuous to implicitly promote ones
name by writing free software and contributing to projects like Apache.
Sorry that I have forgotten to donate this piece of text to the ASF in
the past and instead added such a contemptible hyperlink directing to a
page on www.engelschall.com. I hope this is now fixed and the closed and
holy ASF world is rescued again.
</IRONIC>

<PERSONAL>
I think, I don't have to say that I'm more than angry and disappointed
how developers like me are constantly bashed in the ASF... we can go for
it also in the future, but we should stop looking astonished all the
time if we find out that too less people contribute to the ASF and old
developers like me no longer have a warm feeling here, please. It's our
own fault in thinking that contributions are for free and anonymous just
because our project is a group effort.

IMHO we already have forgotten the golden rule of Open Source
development: if one wants happy and long-term contributing developers
one especially has to make sure they receive the requested credit.
There is an upper limit a project can accept to give, of course. But
credit always has to depend on the amount, quality _and_ duration of
contribution and IMHO cannot be judged by stating that just all people
are equal and so some contributors can be bashed for the fact that their
name occurs more often.

It is correct that my name occurs more often caused by the fact that I
always try to bring in my stuff to the project. But keep in mind that
this is because I _HAVE_ stuff to bring in which I've created _OUTSIDE_
the project. So I think its unfair to bash me just because I try harder
to bring in my additional stuff. If a developer has not much externally
created stuff, he cannot bring it in to the project, of course. But just
because one has more externally created stuff and tries to bring it in,
is IMHO no reason and excuse to bash him for this. It's not my fault
that I write in my freetime more Open Source packages like most of you.

So if you dislike stuff developers want to bring in, decide on the
contribution based on fair technical arguments (pros and cons). But
don't judge the contributions all the time just because you think this
way you "promote" someone (be it RSE, GNU or whoever else). Hell, an
Open Source project is not a group of people to rule their own closed
world and be celebrated in the press for this. It's still an effort to
create the best piece of _software_ money can't buy. So you should stop
thinking about contributors as our enemy. They are the main driving
force of every project, although some people seem to not understand
this at all. And whatever you think about my personal opinion, but
IMHO it's not bad for a project if someone's name is "promoted" with
it, too. What is actually bad are those complains and discussions
which make developers angry and the fact that they result in even less
contributions.
</PERSONAL>


git-svn-id: https://svn.eu.apache.org/repos/asf/httpd/httpd/trunk@84409 13f79535-47bb-0310-9956-ffa450edef68