Fix problem where the Unix mpms had an unitialized variable for child exit status by adding an exit status parameter to ap_wait_all_procs(); with this change, the mpms use ap_wait_all_procs() more like they previously used waitpid().
With the introduction of the exit status parameter, the definition
of ap_wait_t was moved from Apache to APR.  There was some handling
of union wait for the type of the exit status parameter to waitpid()
which I retained (but cannot test).  For WIN32, ap_wait_t was defined
as int (in apr.hw).  No current Windows code uses ap_wait_t, but a
type is required so that references to ap_wait_t in apr_thread_proc.h
can compile on Windows.

Note: There is still a storage leak in the way that the Unix mpms call
ap_wait_all_procs()...  this will be resolved at some later time.


git-svn-id: https://svn.eu.apache.org/repos/asf/httpd/httpd/trunk@85243 13f79535-47bb-0310-9956-ffa450edef68