The following is a patch to fix Dirk's patch from March 14 so that mod_rewrite will work without caveats. Basically Dirk was just dropping patterns of the form ${x} if x didn't resolve to an environment variable. That breaks mod_rewrite since mod_rewrite uses the same syntax for some of its non-environment directives.
The patch below simply puts the behavior for unresolved values back
to what it was before Dirk's patch. If it is unresolved, it is just
left in the directive as is. Further processing will handle the value.

Dirk's optimization still holds if x is resolved. A small price is paid
if x is not resolved, but that shouldn't happen often.

(Note: some tweaks to the comments were negotiated offline after Paul
posted the patch to new-httpd.)
Submitted by:	Paul Reder <rederpj@raleigh.ibm.com>
Reviewed by:	Jeff Trawick


git-svn-id: https://svn.eu.apache.org/repos/asf/httpd/httpd/trunk@85245 13f79535-47bb-0310-9956-ffa450edef68