Change the MaxSpareThreads to 150 for threaded MPM.  This mirrors the new prefork setting and removes the default case where we would spawn three children, but kill two of them immediately because MaxSpareThreads was set to 10 (only one children can be supported with the old value).
Please note that I don't have email access as eBuilt's servers are AWOL
this weekend (none of us have the desire to go in a fix the servers for
semi-obvious reasons).  I'm trying to keep up on posts via apachelabs.org.
So, if you want to change this, just go ahead and commit the changes...


git-svn-id: https://svn.eu.apache.org/repos/asf/httpd/httpd/trunk@90118 13f79535-47bb-0310-9956-ffa450edef68