Lots of small config file changes including:
- Beat everything down to less than 80 character lines.

- Revise the comments for the DSO section to eliminate mention
of ordering problems.

- Add LoadModule lines for all extension modules, commented out.
Note that it is very difficult in the current ./configure
system to make the base modules DSOs, so I didn't include those.

- Remove ServerType comments.

- Add "deny from all" to the root directory container.  The comments
indicate that it should be there.

- Delete a little (in my opinion) excess verbosity in the
description of the .htacess download blocker.

- Remove the CacheNegotiatedDocs directive.  It was wrong,
and is obscure enough that people should probably just look
it up in the docs.

- Make the example for Redirect a little more explicit

- Change the examples and descriptions for READMENAME/HEADERNAME
since, if multiviews are off, they haven't worked as described
since 1.3.something.

- Remove the phf trap example.  This is very old and probably
worthless.

- Remove MetDir/MetSuffix.  These are rarely used.


git-svn-id: https://svn.eu.apache.org/repos/asf/httpd/httpd/trunk@90142 13f79535-47bb-0310-9956-ffa450edef68