With my normal sense of missing the boat :)
This gets the build working on BeOS again :)  Apologies for the delay :(

Jeff changed the order of apr-util and apr to solve a "cleaning" issue but
that makes me uncomfortable as apr-util is dependant on apr, so if we clean
apr-util we shouldn't be altering anything in apr.  If I decide to rebuild
apr-util then apr should still be buildable.  Sorry Jeff but I think we need
a different solution :(

Submitted by:	Peter Schultz <pete@beforever.com>


git-svn-id: https://svn.eu.apache.org/repos/asf/httpd/httpd/trunk@90844 13f79535-47bb-0310-9956-ffa450edef68