We finally got APR working well enough to support buffered reading of .var files, so the .var file issue is gone.
The brokenness of the threaded MPM on FreeBSD is a FreeBSD issue, not
an Apache issue, as far as anyone can tell.  I don't think this should
be in the STATUS file at all, unless somebody thinks we should issue a
big honking warning from configure if somebody tries to enable threaded.


git-svn-id: https://svn.eu.apache.org/repos/asf/httpd/httpd/trunk@90954 13f79535-47bb-0310-9956-ffa450edef68