Get IPv6 vhosts working (again ?).
There were a couple of bogus IPv4-specific comparisons which prevented
the data structures from being built correctly in ap_fini_vhost_config().
One of these (INADDR_ANY) can be easily rewritten to work with IPv6.  The
other (DEFAULT_VHOST_ADDR) isn't necessary for IPv6.  Note that this is
because we don't support the various flavors of _default_ that
Apache-1.3+KAME-patch supports.

PR:      8118


git-svn-id: https://svn.eu.apache.org/repos/asf/httpd/httpd/trunk@92054 13f79535-47bb-0310-9956-ffa450edef68