fix a problem with graceful restart.  If you look at daedalus early in the morning, you will usually see lots ( >50 ) of old generation processes happily serving requests with yesterday's config.  Late in the afternoon, they are usually gone due to MaxRequestsPerChild and perform_idle_server_maintenance.
ap_mpm_pod_signal can time out during connect(), causing it to exit with an
error.  If it was called by ap_mpm_pod_killpg during graceful restart,
ap_mpm_pod_killpg will exit immediately.  This can cause us to not write
enough characters to the pod to kill off all the old generation children.

This fix decouples the loop which writes to the pod from the loop which
generates dummy connect()s.  The connect()s aren't as important as
writing to the pod when there is a steady stream of inbound connections.

also, close the socket and destroy the temporary pool if the connect()
fails.


git-svn-id: https://svn.eu.apache.org/repos/asf/httpd/httpd/trunk@92411 13f79535-47bb-0310-9956-ffa450edef68