Linux 2.4 (MDK 8.0) is fine here as reported earlier.
I'm fine with adjusting that script in a -beta tarball since it isn't a
code change.  However, since official binbuilds are contributed by
committers only, this isn't a big deal as they can be smart enough to
use the updated script anyway.  And, binbuild has some other issues that
will be addressed shortly.  So, even if it doesn't happen, oh well.


git-svn-id: https://svn.eu.apache.org/repos/asf/httpd/httpd/trunk@93189 13f79535-47bb-0310-9956-ffa450edef68