We don't need transfer_brigade() because we already have macros to do that for us, and faster.  :)
Warning, though:  transfer_brigade() was equivalent to APR_BRIGADE_CONCAT(),
and I don't think that was the right behavior for one of the two callers!
I changed that one to use APR_BRIGADE_PREPEND() since by my inspection that
looks like what it should be, but please check me on this.


git-svn-id: https://svn.eu.apache.org/repos/asf/httpd/httpd/trunk@93654 13f79535-47bb-0310-9956-ffa450edef68